# small layout corrections and small performance improvements