… src/Elm/Parser/File.elm
… src/Elm/Parser/Modules.elm
… src/Elm/Parser/Patterns.elm
… src/Elm/Parser/Tokens.elm
… src/Elm/Parser/TypeAnnotation.elm
… src/ParserWithComments.elm
… tests/Elm/Parser/DeclarationsTests.elm
… tests/Elm/Parser/ExpressionTests.elm
… tests/Elm/Parser/PatternTests.elm