@jfmengels 2024-07-30
Do you think you could write tests for the code samples where we were too lenient before this PR? I'd love for the test harness to catch that in the future.