@jfmengels 2024-07-31
Really nice! And the performance improvement is nothing to scoff at either!